% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{sp2df}
\alias{sp2df}
\title{Extract data from PerkinElmer spectrum file}
\usage{
sp2df(path, sampleid = "")
}
\arguments{
\item{path}{path to spectrum file}

\item{sampleid}{string identifying the sample (optional)}
}
\value{
Dataframe with the following columns:
   $ substrateid : chr, included for legacy reasons
   $ sampleid    : chr, see https://github.com/solarchemist/R-common
   $ wavelength  : num, x column from data, wavelength/nm
   $ intensity   : num, y column from data
   $ brand       : chr, instrument metadata
   $ method      : chr, instrument metadata
   $ spectrum    : chr, instrument metadata
   $ ascii       : chr, instrument metadata
   $ peds        : chr, instrument metadata
   $ version     : chr, instrument metadata
   $ filename    : chr, input filename
   $ startdate   : chr, measurement start date
   $ starttime   : chr, measurement start time
   $ enddate     : chr, measurement stop date
   $ endtime     : chr, measurement stop time
   $ xunit       : chr, unit of x-axis (recorded by instrument software)
   $ yunit       : chr, unit of y-axis (recorded by instrument software)
   $ xstart      : chr, start wavelength (PE usually scans from low energy to high, so this is the max wavelength)
   $ xstep       : num, step size/nm
   $ nobs        : num, number of recorded data points
   $ duration    : num, length of measurement in seconds
}
\description{
Reads SP spectrum (ASCII text file) into R dataframe.
Preserves PE parameters (metadata) in the returned dataframe.
Note that x-axis is assumed to be wavelength in nanometres.
}
